Fix tests with decimal

+++ CurrencyConverter.Specs/CurrencySpec.cs
@@ -29,9 +29,9 @@ namespace CurrencyConverter.Specs
         {
             var currency = new Currency("EUR");
             var currencyTarget = new Currency("CHF");
-            new Exchange(currency, currencyTarget, 1.2345);
+            new Exchange(currency, currencyTarget, (decimal)1.2345);
             var othercurrency = new Currency("OTH");
-            new Exchange(othercurrency, currencyTarget, 0.0);
+            new Exchange(othercurrency, currencyTarget, (decimal)0.0);
 
             // After Dijkstra
             currencyTarget.Previous = currency;
