+++ CurrencyConverter.Specs/ExchangeSpec.cs
@@ -11,28 +11,28 @@ namespace CurrencyConverter.Specs
         [Test]
         public void Change_ShouldHaveOriginalCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             Assert.AreEqual(_originalCurrency, change.OriginalCurrency);
         }
 
         [Test]
         public void Change_ShouldHaveTargetCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             Assert.AreEqual(_targetCurrency, change.TargetCurrency);
         }
 
         [Test]
         public void AddCurrencyMapping_ShouldCreateMappingWithOriginalCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             Assert.AreEqual(_originalCurrency.ChangeList[0], change);
         }
 
         [Test]
         public void AddCurrencyMapping_ShouldCreateMappingWithTargetCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             var actual = change.GetNext(_originalCurrency);
             Assert.AreEqual(_targetCurrency, actual);
         }
@@ -40,7 +40,7 @@ namespace CurrencyConverter.Specs
         [Test]
         public void GetNext_ShouldReturnOriginalCurrency_WhenCurrentisTargetCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             var actual = change.GetNext(_targetCurrency);
             Assert.AreEqual(_originalCurrency, actual);
         }
@@ -48,7 +48,7 @@ namespace CurrencyConverter.Specs
         [Test]
         public void GetRateTo_ShouldReturnRate_WithTargetCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             var actual = change.GetRateTo(_originalCurrency);
             Assert.AreEqual(0.8297, actual);
         }
@@ -56,7 +56,7 @@ namespace CurrencyConverter.Specs
         [Test]
         public void GetRateTo_ShouldReturnInvertedRate_WithOriginalCurrency()
         {
-            var change = new Exchange(_originalCurrency, _targetCurrency, 1.2053);
+            var change = new Exchange(_originalCurrency, _targetCurrency, (decimal)1.2053);
             var actual = change.GetRateTo(_targetCurrency);
             Assert.AreEqual(1.2053, actual);
         }
